Move UpdatableFilterLivePageResult to the correct package
